+++ Magic/src/main/java/com/elmakers/mine/bukkit/effect/Visibility.java
@@ -1,0 +1,7 @@
+package com.elmakers.mine.bukkit.effect;
+
+public enum Visibility {
+    ALL,
+    ORIGIN,
+    TARGET
+}